Reestruturação de pacotes OK. TODO: Margin Bottom no Maps. TODO: Opção 'Sair'
