+++ app/src/main/java/com/rogeriocarmo/gnss_mobilecalculator/Model/EpocaGPS.java
@@ -1,4 +1,4 @@
-package Model;
+package com.rogeriocarmo.gnss_mobilecalculator.Model;
 
 import java.util.ArrayList;
 import java.util.Arrays;
