+++ app/src/main/java/com/rogeriocarmo/gnss_mobilecalculator/Model/ResultEpch.java
@@ -1,4 +1,4 @@
-package Model;
+package com.rogeriocarmo.gnss_mobilecalculator.Model;
 
 import java.text.DecimalFormat;
