+++ app/src/main/res/layout/app_bar_side_bar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="View.SideBar">
+    tools:context="com.rogeriocarmo.gnss_mobilecalculator.View.Activity_Main">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
