+++ app/src/main/res/layout/content_side_bar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="View.SideBar"
+    tools:context="com.rogeriocarmo.gnss_mobilecalculator.View.Activity_Main"
     tools:showIn="@layout/app_bar_side_bar">
 
 </android.support.constraint.ConstraintLayout>
