+++ app/src/main/res/layout/fragment_about.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="View.Fragment_About">
+    tools:context="com.rogeriocarmo.gnss_mobilecalculator.View.Fragment_About">
 
     <!-- TODO: Update blank fragment layout -->
