+++ app/src/main/res/raw/log31do10bbbbbbbb.txt
@@ -1,7 +1,7 @@
 # 
 # Header Description:
 # 
-# Version: v2.0.0.1 Platform: 8.0.0 Manufacturer: motorola Model: moto x4
+# Version: v2.0.0.1 Platform: 8.0.0 Manufacturer: motorola com.rogeriocarmo.gnss_mobilecalculator.Model: moto x4
 # 
 # Raw,ElapsedRealtimeMillis,TimeNanos,LeapSecond,TimeUncertaintyNanos,FullBiasNanos,BiasNanos,BiasUncertaintyNanos,DriftNanosPerSecond,DriftUncertaintyNanosPerSecond,HardwareClockDiscontinuityCount,Svid,TimeOffsetNanos,State,ReceivedSvTimeNanos,ReceivedSvTimeUncertaintyNanos,Cn0DbHz,PseudorangeRateMetersPerSecond,PseudorangeRateUncertaintyMetersPerSecond,AccumulatedDeltaRangeState,AccumulatedDeltaRangeMeters,AccumulatedDeltaRangeUncertaintyMeters,CarrierFrequencyHz,CarrierCycles,CarrierPhase,CarrierPhaseUncertainty,MultipathIndicator,SnrInDb,ConstellationType,AgcDb,CarrierFrequencyHz
 #
